requirements.txt
Django
django-crispy-forms
django-debug-toolbar
Pillow
python-decouple
pytz
sqlparse
